First trial for adding conv1dtranspose layer
This attempt ran but did not work in the end. The idea used in this
atttempt was to insert zeros in the input tensor based on kernel size
and apply convolution layer on that. This should have worked but does
not work for some reason. Abandoing this line of thought and
documenting this on this branch.