For transpose, calculating cropped unrolled kernel
Earlier, in conv1d_transpose, the unrolled kernel was always calculated for padding='valid'
case and then the outptu was cropped if padding was 'same'. Now, the unrolled kernel is
calculated in pre-cropped form for padding='same' case so that
conv1d_transpose does not require any cropping operation after matrix multiplication.